refactor(lib): move most things into lib/
All top level *.js files have been moved into lib/. However everything
in bin/ and config/ has been left where it is. Only two minor fixes for
the path to test/local/test-secret.pem were also made.
