+++ bin/server.js
@@ -5,9 +5,9 @@
 var config = require('../config')
 var dbServer = require('fxa-auth-db-server')
 var error = dbServer.errors
-var logger = require('../logging')('bin.server')
-var DB = require('../db/mysql')(logger, error)
-var notifier = require('../notifier.js')(logger, config)
+var logger = require('../lib/logging')('bin.server')
+var DB = require('../lib/db/mysql')(logger, error)
+var notifier = require('../lib/notifier.js')(logger, config)
 
 function shutdown() {
   process.nextTick(process.exit)
