+++ lib/db/patch.js
