+++ lib/db/schema/patch-002-003.sql
