+++ lib/db/schema/patch-003-002.sql
