+++ lib/db/schema/patch-003-004.sql
