+++ lib/db/schema/patch-004-003.sql
