+++ lib/db/schema/patch-005-006.sql
