+++ lib/db/schema/patch-006-005.sql
