+++ lib/db/schema/patch-006-007.sql
