+++ lib/db/schema/patch-007-006.sql
