+++ lib/db/schema/patch-007-008.sql
