+++ lib/db/schema/patch-008-007.sql
