+++ lib/promise.js
