+++ test/local/incorrect-patch-level.js
@@ -4,10 +4,10 @@
 require('ass')
 var dbServer = require('fxa-auth-db-server')
 var log = { trace: console.log, error: console.log, stat: console.log, info: console.log }
-var DB = require('../../db/mysql')(log, dbServer.errors)
+var DB = require('../../lib/db/mysql')(log, dbServer.errors)
 var config = require('../../config')
 var test = require('../ptaptest')
-var patch = require('../../db/patch')
+var patch = require('../../lib/db/patch')
 
 patch.level = 1000000
